+++ forgerock-core/src/androidTest/java/org/forgerock/android/auth/AndroidLSecuredSharedPreferences.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019 ForgeRock. All rights reserved.
+ * Copyright (c) 2019 - 2020 ForgeRock. All rights reserved.
  *
  * This software may be modified and distributed under the terms
  * of the MIT license. See the LICENSE file for details.
@@ -12,12 +12,9 @@ import android.content.Context;
 public class AndroidLSecuredSharedPreferences extends SecuredSharedPreferences {
 
     AndroidLSecuredSharedPreferences(Context context, String fileName, String keyAlias) {
-        super(context, fileName, keyAlias);
+        super(context, fileName, keyAlias, new AndroidLEncryptor(context, keyAlias,
+                new SharedPreferencesSecretKeyStore(keyAlias,
+                        context.getSharedPreferences(fileName, Context.MODE_PRIVATE))));
     }
 
-    @Override
-    protected Encryptor getEncryptor(Context context) {
-        return new AndroidLEncryptor(context, getKeyAlias(),
-                new SharedPreferencesSecretKeyStore(getKeyAlias(), getSharedPreferences()));
-    }
 }